authentication

+++ src/main/java/com/medo/entity/Medicine.java
@@ -1,9 +1,6 @@
 package com.medo.entity;
 
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
+import jakarta.persistence.*;
 import lombok.*;
 import org.springframework.format.annotation.DateTimeFormat;
 
@@ -22,10 +19,12 @@ public class Medicine {
     private long medId;
     private String medicineName;
     private String medicineCompany;
-    private int price;
+    private double price;
     private String medicineType;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate manufacturingDate;
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private LocalDate expiryDate;
+
+
 }
